refactor

+++ iku-mini-program-service/src/main/java/com/iku/sports/mini/admin/service/CourseService.java
@@ -38,5 +38,5 @@ public interface CourseService {
      * @param courseIds the collection of course id
      * @return
      */
-    List<Course> getCourses(final List<Integer> courseIds);
+    List<Course> getCourses(final List<Short> courseIds);
 }
